creating image utils
