+++ deepface/DeepFace.py
@@ -10,7 +10,6 @@ os.environ["TF_USE_LEGACY_KERAS"] = "1"
 # pylint: disable=wrong-import-position
 
 # 3rd party dependencies
-import cv2
 import numpy as np
 import pandas as pd
 import tensorflow as tf
@@ -26,6 +25,7 @@ from deepface.modules import (
     demography,
     detection,
     streaming,
+    preprocessing,
 )
 from deepface import __version__
 
@@ -548,5 +548,5 @@ def detectFace(
     extracted_face = None
     if len(face_objs) > 0:
         extracted_face = face_objs[0]["face"]
-        extracted_face = cv2.resize(extracted_face, target_size)
+        extracted_face = preprocessing.resize_image(img=extracted_face, target_size=target_size)
     return extracted_face
