+++ deepface/modules/recognition.py
@@ -8,10 +8,9 @@ import time
 import numpy as np
 import pandas as pd
 from tqdm import tqdm
-from PIL import Image
 
 # project dependencies
-from deepface.commons import package_utils
+from deepface.commons import package_utils, file_utils
 from deepface.modules import representation, detection, verification
 from deepface.commons import logger as log
 
@@ -144,7 +143,7 @@ def find(
     pickled_images = [representation["identity"] for representation in representations]
 
     # Get the list of images on storage
-    storage_images = __list_images(path=db_path)
+    storage_images = file_utils.list_images(path=db_path)
 
     if len(storage_images) == 0:
         raise ValueError(f"No item found in {db_path}")
@@ -161,7 +160,7 @@ def find(
         if identity in old_images:
             continue
         alpha_hash = current_representation["hash"]
-        beta_hash = package_utils.find_hash_of_file(identity)
+        beta_hash = file_utils.find_hash_of_file(identity)
         if alpha_hash != beta_hash:
             logger.debug(f"Even though {identity} represented before, it's replaced later.")
             replaced_images.append(identity)
@@ -292,31 +291,6 @@ def find(
     return resp_obj
 
 
-def __list_images(path: str) -> List[str]:
-    """
-    List images in a given path
-    Args:
-        path (str): path's location
-    Returns:
-        images (list): list of exact image paths
-    """
-    images = []
-    for r, _, f in os.walk(path):
-        for file in f:
-            exact_path = os.path.join(r, file)
-
-            _, ext = os.path.splitext(exact_path)
-            ext_lower = ext.lower()
-
-            if ext_lower not in {".jpg", ".jpeg", ".png"}:
-                continue
-
-            with Image.open(exact_path) as img:  # lazy
-                if img.format.lower() in ["jpeg", "png"]:
-                    images.append(exact_path)
-    return images
-
-
 def __find_bulk_embeddings(
     employees: List[str],
     model_name: str = "VGG-Face",
@@ -360,7 +334,7 @@ def __find_bulk_embeddings(
         desc="Finding representations",
         disable=silent,
     ):
-        file_hash = package_utils.find_hash_of_file(employee)
+        file_hash = file_utils.find_hash_of_file(employee)
 
         try:
             img_objs = detection.extract_faces(
